voice_to_image/requirements.txt
streamlit
openai
python-dotenv
requests
audio-recorder-streamlit

